+++ CHANGES.rst
@@ -8,6 +8,7 @@ New Features:
 
 * Sphinx frontend: use the `today` and `today_fmt` configuration variables for
   the date on the title page
+* Sphinx frontend: allow extensions access to the builder object (issue #155)
 
 Fixed: